myproject/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.1
Django==5.1
django-allauth==64.2.0
django-widget-tweaks==1.5.0
gunicorn==23.0.0
idna==3.8
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
whitenoise==6.7.0
